+++ rust/tvm-ffi/src/extra/mod.rs
@@ -16,5 +16,6 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+pub mod dispatch;
 pub mod module;
 pub mod structural_visit;
